Sharing option for mangas and chapters, confirm before deleting all chapters

+++ iOS/UI/Manga/MangaViewController.swift
@@ -313,18 +313,45 @@ class MangaViewController: UIViewController {
                 })
             }
 
-            if DownloadManager.shared.hasDownloadedChapter(for: manga) {
-                subActions.append(UIAction(
-                    title: NSLocalizedString("REMOVE_ALL_DOWNLOADS", comment: ""),
-                    image: UIImage(systemName: "trash"),
-                    attributes: .destructive
-                ) { [weak self] _ in
-                    guard let self = self else { return }
-                    DownloadManager.shared.deleteChapters(for: self.manga)
-                })
+            subMenus.append(UIMenu(title: "", options: .displayInline, children: subActions))
+
+            if URL(string: manga.url ?? "") != nil {
+                subMenus.append(UIMenu(title: "", options: .displayInline, children: [
+                    UIAction(
+                        title: NSLocalizedString("SHARE", comment: ""),
+                        image: UIImage(systemName: "square.and.arrow.up")
+                    ) { [weak self] _ in
+                        guard let self = self else { return }
+                        guard let url = URL(string: self.manga.url ?? "") else { return }
+
+                        let activityViewController = UIActivityViewController(
+                            activityItems: [url],
+                            applicationActivities: nil
+                        )
+                        activityViewController.popoverPresentationController?.sourceView = self.view
+
+                        self.present(activityViewController, animated: true)
+                    }
+                ]))
             }
 
-            subMenus.append(UIMenu(title: "", options: .displayInline, children: subActions))
+            if DownloadManager.shared.hasDownloadedChapter(for: manga) {
+                subMenus.append(UIMenu(title: "", options: .displayInline, children: [
+                    UIAction(
+                        title: NSLocalizedString("REMOVE_ALL_DOWNLOADS", comment: ""),
+                        image: UIImage(systemName: "trash"),
+                        attributes: .destructive
+                    ) { [weak self] _ in
+                        guard let self = self else { return }
+                        confirmAction(title: NSLocalizedString("REMOVE_ALL_DOWNLOADS", comment: ""),
+                                      message: NSLocalizedString("REMOVE_ALL_DOWNLOADS_CONFIRM", comment: ""),
+                                      continueActionName: NSLocalizedString("REMOVE", comment: "")) { [weak self] in
+                            guard let self = self else { return }
+                            DownloadManager.shared.deleteChapters(for: self.manga)
+                        }
+                    }
+                ]))
+            }
 
             let menu = UIMenu(title: "", children: subMenus)
 
@@ -477,18 +504,32 @@ class MangaViewController: UIViewController {
     @objc func deleteSelectedChapters() {
         guard let selected = tableView.indexPathsForSelectedRows else { return }
 
-        let alertView = UIAlertController(
-            title: NSLocalizedString("REMOVE_DOWNLOADS", comment: ""),
-            message: NSLocalizedString("REMOVE_DOWNLOADS_CONFIRM", comment: ""),
-            preferredStyle: UIDevice.current.userInterfaceIdiom == .pad ? .alert : .actionSheet
-        )
-        alertView.addAction(UIAlertAction(title: NSLocalizedString("REMOVE", comment: ""), style: .destructive) { _ in
+        confirmAction(title: NSLocalizedString("REMOVE_DOWNLOADS", comment: ""),
+                      message: NSLocalizedString("REMOVE_DOWNLOADS_CONFIRM", comment: ""),
+                      continueActionName: NSLocalizedString("REMOVE", comment: "")) {
             DownloadManager.shared.delete(chapters: selected.map { self.sortedChapters[$0.row] })
             self.setEditing(false, animated: true)
-        })
+        }
+    }
+
+    func confirmAction(
+        title: String,
+        message: String,
+        continueActionName: String = NSLocalizedString("CONTINUE", comment: ""),
+        destructive: Bool = true,
+        proceed: @escaping () -> Void
+    ) {
+        let alertView = UIAlertController(
+            title: title,
+            message: message,
+            preferredStyle: UIDevice.current.userInterfaceIdiom == .pad ? .alert : .actionSheet
+        )
+
+        let action = UIAlertAction(title: continueActionName, style: destructive ? .destructive : .default) { _ in proceed() }
+        alertView.addAction(action)
+
         alertView.addAction(UIAlertAction(title: NSLocalizedString("CANCEL", comment: ""), style: .cancel))
         present(alertView, animated: true)
-
     }
 }
 
@@ -665,6 +706,7 @@ extension MangaViewController: UITableViewDataSource {
         UIContextMenuConfiguration(identifier: nil, previewProvider: nil) { [weak self] _ -> UIMenu? in
             guard let self = self else { return nil }
             var actions: [UIMenuElement] = []
+
             // download action
             let downloadAction: UIMenuElement
             let downloadStatus = DownloadManager.shared.getDownloadStatus(for: self.sortedChapters[indexPath.row])
@@ -691,6 +733,7 @@ extension MangaViewController: UITableViewDataSource {
                 }
             }
             actions.append(UIMenu(title: "", options: .displayInline, children: [downloadAction]))
+
             // marking actions
             let history = self.readHistory[self.sortedChapters[indexPath.row].id] ?? (0, 0)
             if history.1 <= 0 || history.0 > 0 {
@@ -731,6 +774,24 @@ extension MangaViewController: UITableViewDataSource {
                     }
                 ])
                 actions.append(previousSubmenu)
+            }
+
+            // sharing action
+            if URL(string: self.sortedChapters[indexPath.row].url ?? "") != nil {
+                actions.append(UIMenu(title: "", options: .displayInline, children: [
+                    UIAction(title: NSLocalizedString("SHARE", comment: ""), image: nil) { [weak self] _ in
+                        guard let self = self else { return }
+                        guard let url = URL(string: self.sortedChapters[indexPath.row].url ?? "") else { return }
+
+                        let activityViewController = UIActivityViewController(
+                            activityItems: [url],
+                            applicationActivities: nil
+                        )
+                        activityViewController.popoverPresentationController?.sourceView = self.view
+
+                        self.present(activityViewController, animated: true)
+                    }
+                ]))
             }
             return UIMenu(title: "", children: actions)
         }
